Restructure test folder
reflect test folder with lib folder for easier maintenance
