+++ test/services/location_search_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import './test_app_widget.dart';
+import '../test_app_widget.dart';
 
 void main() {
   group('Location Search', () {
