+++ test/widgets/custom_appbar_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import './test_app_widget.dart';
+import '../test_app_widget.dart';
 
 void main() {
   group('Custom Appbar', () {
