+++ test/widgets/custom_drawer_test.dart
@@ -1,7 +1,7 @@
 import 'package:concordia_navigation/widgets/custom_drawer.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import './test_app_widget.dart';
+import '../test_app_widget.dart';
 
 void main() {
   group('Custom Drawer', () {
