+++ test/widgets/map_widget_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import './test_app_widget.dart';
+import '../test_app_widget.dart';
 
 void main() {
   group('MapWidget', () {
